[tidy] split basic test into basic, serialize, tree
