+++ test/CMakeLists.txt
@@ -58,6 +58,8 @@ endfunction()
 ryml_add_test(callbacks)
 ryml_add_test(stack)
 ryml_add_test(parser)
+ryml_add_test(tree)
+ryml_add_test(serialize)
 ryml_add_test(basic)
 ryml_add_test(basic_json)
 ryml_add_test(preprocess)
